+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.3
+
+* Added `keepRemind` parameter to disable the default behavior (default behavior is auto requests for the review on each new version), and it's `true` by default.
+* `openStore` method will open the `fallbackUrl` if available on not supported platforms.
+
 ## 0.0.2
 
 * Add `openStore` to open the current app on the store.